requirements.txt
fastapi
uvicorn
numpy
pillow
tensorflow
openai
python-dotenv
elevenlabs
pydub
